requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
fonttools==4.51.0
frozenlist==1.4.1
idna==3.7
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.1
magic-filter==1.0.12
matplotlib==3.8.4
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prompt-toolkit==3.0.43
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rauth==0.7.3
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
